+++ pkg/hmenum/rega_script.go
@@ -26,6 +26,7 @@ const (
 	RegaScriptGetSystemVariableDescriptions RegaScript = "get_system_variable_descriptions"
 	RegaScriptSetProgramState               RegaScript = "set_program_state"
 	RegaScriptExecuteProgramConditional     RegaScript = "execute_program_conditional"
+	RegaScriptDeleteProgram                 RegaScript = "delete_program"
 	RegaScriptSetSystemVariable             RegaScript = "set_system_variable"
 	// Lifecycle scripts. Names match
 	// Reference where one exists
@@ -69,6 +70,7 @@ var AllRegaScripts = []RegaScript{
 	RegaScriptGetSystemVariableDescriptions,
 	RegaScriptSetProgramState,
 	RegaScriptExecuteProgramConditional,
+	RegaScriptDeleteProgram,
 	RegaScriptSetSystemVariable,
 	RegaScriptCreateSystemVariable,
 	RegaScriptUpdateSystemVariable,